Added requirements.txt file for easy package installation
Moved the button class to a separate file

started moving clipboard related functions to a separate file

Implemented dotenv

Initial readme.md

Started moving SFTP operations

Corrected import for parse_case_number

OS operations moved

renamed folders, corrected imports accordingly

Moving the GUI initialization to a separate file

Moved watchdog to a separate file
