+++ README.md
@@ -1,0 +1,24 @@
+# TAC Survival Toolkit
+
+## Installation
+
+1. Clone the repository to your local machine.
+
+```
+git clone git@github.com:StamLord/TAC-Survival-Toolkit.git
+```
+
+2. Install the dependencies.
+
+It is recommended to use a virtual environment.
+
+```
+cd TAC-Survival-Toolkit
+python3 -m venv env
+.\env\Scripts\activate
+pip install -r requirements.txt
+```
+
+3. Configure the env varilables.  
+   This is done by creating a file called `.env` in the root directory of the repository.  
+   Please contact the repository owner if you do not have access to this file.
